refactor: 360도 회전 컴포넌트 위치 변경
